Split token counting to separate file
